tried to format code to run like the gymnasium library
